Some methods implemented
