+++ entities/tasks/src/lib.rs
@@ -5,20 +5,20 @@ pub mod repo;
 use yatir_core::prelude::*;
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct ToDo {
+pub struct Task {
     pub parent: Index,
     pub id: Index,
     pub name: String,
     pub description: String,
 }
 
-pub(crate) fn default_todo_name() -> String {
+pub(crate) fn default_task_name() -> String {
     "New folder".to_string()
 }
 
 pub mod prelude {
     pub use super::{
-        repo::{CreateToDo, ToDoRepo, UpdateToDo},
-        ToDo,
+        repo::{CreateTask, TaskRepo, UpdateTask},
+        Task,
     };
 }
